requirements.txt
asgiref==3.3.4
braintree==4.9.0
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
